Rename GPU device array helpers from cuda_ to gpu_
Common utility helpers between CUDA and ROCm are renamed:

- core/kernels/cuda_device_array.h -> core/kernels/gpu_device_array.h
- core/kernels/cuda_device_array_gpu.h -> core/kernels/gpu_device_array_gpu.h

In subsequent PRs, ROCm-specific terminology changes would be introduced.
