+++ tensorflow/core/kernels/split_lib_gpu.cu.cc
@@ -23,7 +23,7 @@ limitations under the License.
 
 #include "tensorflow/core/framework/register_types.h"
 #include "tensorflow/core/framework/tensor_types.h"
-#include "tensorflow/core/kernels/cuda_device_array_gpu.h"
+#include "tensorflow/core/kernels/gpu_device_array_gpu.h"
 #include "tensorflow/core/kernels/split_lib_gpu.h"
 #include "tensorflow/core/util/cuda_kernel_helper.h"
